refactor(pglite): promote to prisma submodule
Make pglite part of the core codebase, so that we can use it in
environments where postgres may not be available, such as in
browser-only environments like StackBlitz.

- Rename prisma -> prisma/index to host submodules
- Move pglite-related logic to pglite
  - Wrap actual init of PGlite as factory, and not
    within module load
- Rework vitest.setup to depend on new submodule
  - Keep resetDb, since that is only useful to testing
